# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroAdmin platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ServerStatus = typeof ServerStatus[keyof typeof ServerStatus];


export const ServerStatus = {
  online: 'online',
  offline: 'offline',
  warning: 'warning',
  unknown: 'unknown',
} as const;

export interface Server {
  id: number;
  name: string;
  host: string;
  /** @nullable */
  description?: string | null;
  status: ServerStatus;
  /** @nullable */
  agentToken?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface ServerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  host: string;
  description?: string;
}

export type ServerUpdateStatus = typeof ServerUpdateStatus[keyof typeof ServerUpdateStatus];


export const ServerUpdateStatus = {
  online: 'online',
  offline: 'offline',
  warning: 'warning',
  unknown: 'unknown',
} as const;

export interface ServerUpdate {
  /** @minLength 1 */
  name?: string;
  host?: string;
  description?: string;
  status?: ServerUpdateStatus;
}

export interface ServerMetrics {
  id: number;
  serverId: number;
  cpuPercent: number;
  memoryUsedMb: number;
  memoryTotalMb: number;
  diskUsedGb: number;
  diskTotalGb: number;
  uptimeSeconds: number;
  /** @nullable */
  networkRxMbps?: number | null;
  /** @nullable */
  networkTxMbps?: number | null;
  /** @nullable */
  topProcess?: string | null;
  recordedAt: string;
}

export interface MetricsInput {
  cpuPercent: number;
  memoryUsedMb: number;
  memoryTotalMb: number;
  diskUsedGb: number;
  diskTotalGb: number;
  uptimeSeconds: number;
  networkRxMbps?: number;
  networkTxMbps?: number;
  topProcess?: string;
  lastLogs?: string;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  cpu_high: 'cpu_high',
  memory_high: 'memory_high',
  disk_high: 'disk_high',
  server_down: 'server_down',
  custom: 'custom',
} as const;

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  info: 'info',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface Alert {
  id: number;
  serverId: number;
  type: AlertType;
  message: string;
  severity: AlertSeverity;
  resolved: boolean;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  aiAnalysis?: string | null;
  /** @nullable */
  lastLogs?: string | null;
  createdAt: string;
}

export interface AiAnalysis {
  alertId: number;
  analysis: string;
}

export interface DashboardSummary {
  totalServers: number;
  onlineServers: number;
  offlineServers: number;
  warningServers: number;
  activeAlerts: number;
  criticalAlerts: number;
}