# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroAdmin platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';
import type { AlertType } from './alertType';

export interface Alert {
  id: number;
  serverId: number;
  type: AlertType;
  message: string;
  severity: AlertSeverity;
  resolved: boolean;
  /** @nullable */
  resolvedAt?: Date | null;
  /** @nullable */
  aiAnalysis?: string | null;
  /** @nullable */
  lastLogs?: string | null;
  createdAt: Date;
}