# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroAdmin platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalServers: number;
  onlineServers: number;
  offlineServers: number;
  warningServers: number;
  activeAlerts: number;
  criticalAlerts: number;
}