# lib/api-zod/src/generated/types/metricsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroAdmin platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MetricsInput {
  cpuPercent: number;
  memoryUsedMb: number;
  memoryTotalMb: number;
  diskUsedGb: number;
  diskTotalGb: number;
  uptimeSeconds: number;
  networkRxMbps?: number;
  networkTxMbps?: number;
  topProcess?: string;
  lastLogs?: string;
}