# lib/api-zod/src/generated/types/server.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroAdmin platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ServerStatus } from './serverStatus';

export interface Server {
  id: number;
  name: string;
  host: string;
  /** @nullable */
  description?: string | null;
  status: ServerStatus;
  /** @nullable */
  agentToken?: string | null;
  createdAt: Date;
  updatedAt?: Date;
}