# lib/api-zod/src/generated/types/serverInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroAdmin platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ServerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  host: string;
  description?: string;
}