# lib/api-zod/src/generated/types/serverMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroAdmin platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ServerMetrics {
  id: number;
  serverId: number;
  cpuPercent: number;
  memoryUsedMb: number;
  memoryTotalMb: number;
  diskUsedGb: number;
  diskTotalGb: number;
  uptimeSeconds: number;
  /** @nullable */
  networkRxMbps?: number | null;
  /** @nullable */
  networkTxMbps?: number | null;
  /** @nullable */
  topProcess?: string | null;
  recordedAt: Date;
}