# lib/api-zod/src/generated/types/serverUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroAdmin platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ServerUpdateStatus } from './serverUpdateStatus';

export interface ServerUpdate {
  /** @minLength 1 */
  name?: string;
  host?: string;
  description?: string;
  status?: ServerUpdateStatus;
}